[Chore] Move the asset-execution tests into the modules they test
The facade was still holding all 26 tests, so it remained a 1 500-line
file and no test sat beside its subject. Each now lives with what it
covers: cache validation and durability in `cache`, deobfuscation and
asset-info decryption in `crypto`, filters and task construction in
`planning`, the 3D pipeline's eight cases in `haruki_3d`, cookies and
HTTP retry in `provider`, backlog sizing in `runner`.

`test_region`, the AES fixtures and `encrypt_asset_info` move to
`test_support` rather than being copied. The AES pair matters: two
modules each carrying their own key and IV would drift apart with nothing
to catch it, and `encrypt_asset_info` is needed by both the provider
tests that serve an encrypted document and the download tests that fetch
one.

Same 26 tests before and after. `asset_execution.rs` is now 25 lines.
