+++ src/core/asset_execution/crypto.rs
@@ -86,3 +86,81 @@ pub(super) type Aes256CbcDec = cbc::Decryptor<aes::Aes256>;
 pub(super) type Aes192CbcDec = cbc::Decryptor<aes::Aes192>;
 
 pub(super) type Aes128CbcDec = cbc::Decryptor<aes::Aes128>;
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashMap;
+
+    use super::super::crypto::{decrypt_asset_bundle_info, deobfuscate, deobfuscate_owned};
+
+    use super::super::model::{AssetBundleDetail, AssetBundleInfo, AssetCategory};
+
+    use super::super::test_support::{encrypt_asset_info, TEST_AES_IV_HEX, TEST_AES_KEY_HEX};
+
+    #[test]
+    fn decrypt_asset_info_round_trips_msgpack_payload() {
+        let info = AssetBundleInfo {
+            version: Some("1".to_string()),
+            os: Some("ios".to_string()),
+            bundles: HashMap::from([(
+                "start/a".to_string(),
+                AssetBundleDetail {
+                    bundle_name: "start/a".to_string(),
+                    cache_file_name: "a".to_string(),
+                    cache_directory_name: "d".to_string(),
+                    hash: "hash".to_string(),
+                    category: AssetCategory::StartApp,
+                    crc: 123,
+                    file_size: 1,
+                    dependencies: Vec::new(),
+                    paths: Vec::new(),
+                    is_builtin: false,
+                    is_relocate: None,
+                    md5_hash: None,
+                    download_path: None,
+                },
+            )]),
+        };
+
+        let encrypted = encrypt_asset_info(&info);
+        let decrypted =
+            decrypt_asset_bundle_info(TEST_AES_KEY_HEX, TEST_AES_IV_HEX, &encrypted).unwrap();
+        assert_eq!(decrypted.version.as_deref(), Some("1"));
+        assert!(decrypted.bundles.contains_key("start/a"));
+    }
+
+    #[test]
+    fn deobfuscate_matches_go_headers() {
+        assert_eq!(
+            deobfuscate(&[0x20, 0x00, 0x00, 0x00, 1, 2, 3]),
+            vec![1, 2, 3]
+        );
+        assert_eq!(deobfuscate(&[9, 8, 7]), vec![9, 8, 7]);
+    }
+
+    #[test]
+    fn deobfuscate_owned_reuses_the_input_allocation() {
+        let simple = vec![0x20, 0x00, 0x00, 0x00, 1, 2, 3];
+        let simple_pointer = simple.as_ptr();
+        let simple_capacity = simple.capacity();
+        let simple = deobfuscate_owned(simple);
+        assert_eq!(simple, vec![1, 2, 3]);
+        assert_eq!(simple.as_ptr(), simple_pointer);
+        assert_eq!(simple.capacity(), simple_capacity);
+
+        let pattern = [0xff, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00];
+        let decoded = (0..132).map(|index| index as u8).collect::<Vec<_>>();
+        let mut xor = vec![0x10, 0x00, 0x00, 0x00];
+        xor.extend(decoded.iter().enumerate().map(|(index, byte)| {
+            if index < 128 {
+                byte ^ pattern[index % pattern.len()]
+            } else {
+                *byte
+            }
+        }));
+        let xor_pointer = xor.as_ptr();
+        let xor = deobfuscate_owned(xor);
+        assert_eq!(xor, decoded);
+        assert_eq!(xor.as_ptr(), xor_pointer);
+    }
+}
