+++ src/core/asset_execution/planning.rs
@@ -234,3 +234,353 @@ impl AssetExecutionContext {
         tasks
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::collections::{BTreeMap, HashMap};
+    use std::path::Path;
+
+    use tempfile::tempdir;
+
+    use crate::core::config::{AppConfig, RawBundleExportConfig, RegionProviderConfig};
+    use crate::core::download_records::DownloadRecord;
+    use crate::core::models::{AssetUpdateMode, AssetUpdateRequest};
+
+    use super::super::model::{
+        AssetBundleDetail, AssetBundleInfo, AssetCategory, AssetExecutionContext,
+    };
+    use super::super::planning::{raw_bundle_output_path, should_download_bundle};
+
+    use super::super::test_support::test_region;
+
+    #[test]
+    fn raw_bundle_filters_are_independent_of_haruki_3d() {
+        let mut region = test_region(RegionProviderConfig::ColorfulPalette {
+            asset_info_url_template: "https://example.com/info".to_string(),
+            asset_bundle_url_template: "https://example.com/{bundle_path}".to_string(),
+            profile: "production".to_string(),
+            profile_hashes: BTreeMap::new(),
+            required_cookies: false,
+            cookie_bootstrap_url: None,
+        });
+        region.export.raw_bundles = Some(RawBundleExportConfig {
+            output_dir: None,
+            include: vec!["^live_pv/model/characterv2/body/".to_string()],
+            exclude: Vec::new(),
+        });
+        region.filters.on_demand.clear();
+        region.filters.skip = vec![".*".to_string()];
+        let request = AssetUpdateRequest {
+            region: "jp".to_string(),
+            asset_version: Some("1".to_string()),
+            asset_hash: Some("hash".to_string()),
+            dry_run: false,
+            mode: AssetUpdateMode::Update,
+        };
+        let config = AppConfig::default();
+
+        let executor = AssetExecutionContext::new(&config, "jp", &region, &request).unwrap();
+        assert!(
+            executor.matches_raw_bundle_filters("live_pv/model/characterv2/body/01"),
+            "raw bundle retention must remain independent while 3D is disabled"
+        );
+        assert!(!executor.matches_raw_bundle_filters("live_pv/model/characterv2/face/01"));
+
+        let detail = |bundle_name: &str| AssetBundleDetail {
+            bundle_name: bundle_name.to_string(),
+            cache_file_name: String::new(),
+            cache_directory_name: String::new(),
+            hash: format!("{bundle_name}-hash"),
+            category: AssetCategory::OnDemand,
+            crc: 0,
+            file_size: 1,
+            dependencies: Vec::new(),
+            paths: Vec::new(),
+            is_builtin: false,
+            is_relocate: None,
+            md5_hash: None,
+            download_path: None,
+        };
+        let info = AssetBundleInfo {
+            version: Some("1".to_string()),
+            os: Some("ios".to_string()),
+            bundles: HashMap::from([
+                (
+                    "live_pv/model/characterv2/body/01".to_string(),
+                    detail("live_pv/model/characterv2/body/01"),
+                ),
+                (
+                    "live_pv/model/characterv2/face/01".to_string(),
+                    detail("live_pv/model/characterv2/face/01"),
+                ),
+            ]),
+        };
+        let tasks = executor.build_raw_bundle_filter_tasks(&info);
+        assert_eq!(tasks.len(), 1);
+        assert_eq!(tasks[0].bundle_path, "live_pv/model/characterv2/body/01");
+    }
+
+    #[test]
+    fn build_download_tasks_skips_unchanged_and_queues_changed() {
+        let region = test_region(RegionProviderConfig::ColorfulPalette {
+            asset_info_url_template: String::new(),
+            asset_bundle_url_template: String::new(),
+            profile: "production".to_string(),
+            profile_hashes: BTreeMap::from([("production".to_string(), "abc".to_string())]),
+            required_cookies: false,
+            cookie_bootstrap_url: None,
+        });
+        let config = AppConfig::default();
+        let request = AssetUpdateRequest {
+            region: "jp".to_string(),
+            asset_version: Some("1".to_string()),
+            asset_hash: Some("hash".to_string()),
+            dry_run: false,
+            mode: AssetUpdateMode::Update,
+        };
+        let ctx = AssetExecutionContext::new(&config, "jp", &region, &request).unwrap();
+
+        let detail = |hash: &str| AssetBundleDetail {
+            bundle_name: String::new(),
+            cache_file_name: String::new(),
+            cache_directory_name: String::new(),
+            hash: hash.to_string(),
+            category: AssetCategory::StartApp,
+            crc: 0,
+            file_size: 1,
+            dependencies: Vec::new(),
+            paths: Vec::new(),
+            is_builtin: false,
+            is_relocate: None,
+            md5_hash: None,
+            download_path: None,
+        };
+        let info = AssetBundleInfo {
+            version: Some("1".to_string()),
+            os: Some("ios".to_string()),
+            bundles: HashMap::from([
+                ("start/a".to_string(), detail("h1")),
+                ("start/aa".to_string(), detail("h2")),
+            ]),
+        };
+
+        // Recorded hash matches -> skipped; bundle absent from record -> queued.
+        let record = DownloadRecord::from([("start/a".to_string(), "h1".to_string())]);
+        let tasks = ctx
+            .build_download_tasks(&info, &record, &DownloadRecord::new(), false)
+            .unwrap();
+        let paths: Vec<&str> = tasks.iter().map(|task| task.bundle_path.as_str()).collect();
+        assert!(
+            !paths.contains(&"start/a"),
+            "unchanged bundle must be skipped"
+        );
+        assert!(paths.contains(&"start/aa"), "new bundle must be queued");
+
+        // Recorded hash differs -> re-queued.
+        let stale = DownloadRecord::from([("start/a".to_string(), "OLD".to_string())]);
+        let tasks = ctx
+            .build_download_tasks(&info, &stale, &DownloadRecord::new(), false)
+            .unwrap();
+        let paths: Vec<&str> = tasks.iter().map(|task| task.bundle_path.as_str()).collect();
+        assert!(
+            paths.contains(&"start/a"),
+            "changed bundle must be re-queued"
+        );
+    }
+
+    #[tokio::test]
+    async fn build_download_tasks_routes_3d_only_matches_to_staging() {
+        let temp = tempdir().unwrap();
+        let mut region = test_region(RegionProviderConfig::ColorfulPalette {
+            asset_info_url_template: String::new(),
+            asset_bundle_url_template: String::new(),
+            profile: "production".to_string(),
+            profile_hashes: BTreeMap::from([("production".to_string(), "abc".to_string())]),
+            required_cookies: false,
+            cookie_bootstrap_url: None,
+        });
+        region.filters.on_demand.clear();
+        region.export.haruki_3d = crate::core::config::Haruki3dExportConfig {
+            enabled: true,
+            work_dir: temp.path().join("3d-work").to_string_lossy().into_owned(),
+            manifest_file: temp
+                .path()
+                .join("runtime/haruki-3d-export-manifest.json")
+                .to_string_lossy()
+                .into_owned(),
+            include: vec!["^(start/a|live_pv/model/characterv2/body/)".to_string()],
+            ..crate::core::config::Haruki3dExportConfig::default()
+        };
+        let config = AppConfig::default();
+        let request = AssetUpdateRequest {
+            region: "jp".to_string(),
+            asset_version: Some("1".to_string()),
+            asset_hash: Some("hash".to_string()),
+            dry_run: false,
+            mode: AssetUpdateMode::Update,
+        };
+        let executor = AssetExecutionContext::new(&config, "jp", &region, &request).unwrap();
+        let detail = |bundle_name: &str, category| AssetBundleDetail {
+            bundle_name: bundle_name.to_string(),
+            cache_file_name: String::new(),
+            cache_directory_name: String::new(),
+            hash: format!("{bundle_name}-hash"),
+            category,
+            crc: 0,
+            file_size: 1,
+            dependencies: Vec::new(),
+            paths: Vec::new(),
+            is_builtin: false,
+            is_relocate: None,
+            md5_hash: None,
+            download_path: None,
+        };
+        let info = AssetBundleInfo {
+            version: Some("1".to_string()),
+            os: Some("ios".to_string()),
+            bundles: HashMap::from([
+                (
+                    "start/a".to_string(),
+                    detail("start/a", AssetCategory::StartApp),
+                ),
+                (
+                    "live_pv/model/characterv2/body/01".to_string(),
+                    detail("live_pv/model/characterv2/body/01", AssetCategory::OnDemand),
+                ),
+            ]),
+        };
+
+        let tasks = executor
+            .build_download_tasks(&info, &DownloadRecord::new(), &DownloadRecord::new(), false)
+            .unwrap();
+        let paths: Vec<&str> = tasks.iter().map(|task| task.bundle_path.as_str()).collect();
+        assert_eq!(
+            paths,
+            vec!["start/a", "live_pv/model/characterv2/body/01"],
+            "3D matches with missing staging must be merged once after ordinary download filtering"
+        );
+        assert!(
+            tasks[0].export_payloads && tasks[0].stage_haruki_3d,
+            "ordinary tasks must export payloads"
+        );
+        assert!(
+            !tasks[1].export_payloads && tasks[1].stage_haruki_3d,
+            "3D-only tasks must only stage raw bundles"
+        );
+
+        let haruki_3d_record = DownloadRecord::from([(
+            "live_pv/model/characterv2/body/01".to_string(),
+            "live_pv/model/characterv2/body/01-hash".to_string(),
+        )]);
+        let tasks = executor
+            .build_download_tasks(
+                &info,
+                &DownloadRecord::new(),
+                &haruki_3d_record,
+                executor.can_reuse_haruki_3d_download_record().await,
+            )
+            .unwrap();
+        assert_eq!(
+            tasks
+                .iter()
+                .map(|task| task.bundle_path.as_str())
+                .collect::<Vec<_>>(),
+            vec!["start/a", "live_pv/model/characterv2/body/01"],
+            "the independent 3D record must not skip bundles when the runtime manifest is missing"
+        );
+
+        let manifest = Path::new(&region.export.haruki_3d.manifest_file);
+        std::fs::create_dir_all(manifest.parent().unwrap()).unwrap();
+        std::fs::write(manifest, b"{broken").unwrap();
+        let tasks = executor
+            .build_download_tasks(
+                &info,
+                &DownloadRecord::new(),
+                &haruki_3d_record,
+                executor.can_reuse_haruki_3d_download_record().await,
+            )
+            .unwrap();
+        assert_eq!(
+            tasks
+                .iter()
+                .map(|task| task.bundle_path.as_str())
+                .collect::<Vec<_>>(),
+            vec!["start/a", "live_pv/model/characterv2/body/01"],
+            "a malformed 3D runtime manifest must not make the download record reusable"
+        );
+
+        std::fs::write(manifest, br#"{"parts/example":{"bundleLength":1}}"#).unwrap();
+        let tasks = executor
+            .build_download_tasks(
+                &info,
+                &DownloadRecord::new(),
+                &haruki_3d_record,
+                executor.can_reuse_haruki_3d_download_record().await,
+            )
+            .unwrap();
+        assert_eq!(
+            tasks
+                .iter()
+                .map(|task| task.bundle_path.as_str())
+                .collect::<Vec<_>>(),
+            vec!["start/a"],
+            "the independent 3D record must skip an unchanged bundle even after staging cleanup"
+        );
+    }
+
+    #[test]
+    fn download_filters_match_go_logic() {
+        let region = test_region(RegionProviderConfig::ColorfulPalette {
+            asset_info_url_template: "".to_string(),
+            asset_bundle_url_template: "".to_string(),
+            profile: "production".to_string(),
+            profile_hashes: BTreeMap::from([("production".to_string(), "abc".to_string())]),
+            required_cookies: false,
+            cookie_bootstrap_url: None,
+        });
+
+        assert!(should_download_bundle(
+            &region,
+            "start/a",
+            &AssetCategory::StartApp
+        ));
+        assert!(should_download_bundle(
+            &region,
+            "ond/a",
+            &AssetCategory::OnDemand
+        ));
+        assert!(should_download_bundle(
+            &region,
+            "live_pv/model/characterv2/body/99/0018/ladies_s",
+            &AssetCategory::LivePv
+        ));
+        assert!(!should_download_bundle(
+            &region,
+            "other/a",
+            &AssetCategory::OnDemand
+        ));
+        assert!(!should_download_bundle(
+            &region,
+            "character/member/001",
+            &AssetCategory::LivePv
+        ));
+    }
+
+    #[test]
+    fn raw_bundle_output_path_appends_bundle_extension_and_rejects_unsafe_paths() {
+        let root = std::path::Path::new("/tmp/raw-root");
+        assert_eq!(
+            raw_bundle_output_path(root, "live_pv/model/character/body/foo").unwrap(),
+            root.join("live_pv/model/character/body/foo.bundle")
+        );
+        assert_eq!(
+            raw_bundle_output_path(root, "character/motion/costume_setting/01_00.bundle").unwrap(),
+            root.join("character/motion/costume_setting/01_00.bundle")
+        );
+        assert!(raw_bundle_output_path(root, "").is_err());
+        assert!(raw_bundle_output_path(root, "/absolute/path").is_err());
+        assert!(raw_bundle_output_path(root, "../escape").is_err());
+        assert!(raw_bundle_output_path(root, "safe/../escape").is_err());
+        assert!(raw_bundle_output_path(root, "safe/./escape").is_err());
+    }
+}
